+++ database/migrations/mikoyan/src/convert/exam_results.rs
@@ -31,9 +31,9 @@ impl<'de> serde::de::Visitor<'de> for ExamResultsVisitor {
                     }
 
                     exam_time_in_seconds = match map.next_value()? {
-                        Bson::Int32(v) => Some(v),
-                        Bson::Int64(v) => Some(v as i32),
-                        Bson::Double(v) => Some(v as i32),
+                        Bson::Int32(v) => Some(v as u32),
+                        Bson::Int64(v) => Some(v as u32),
+                        Bson::Double(v) => Some(v as u32),
                         _ => None,
                     };
                 }
@@ -43,9 +43,9 @@ impl<'de> serde::de::Visitor<'de> for ExamResultsVisitor {
                     }
 
                     number_of_correct_answers = match map.next_value()? {
-                        Bson::Int32(v) => Some(v),
-                        Bson::Int64(v) => Some(v as i32),
-                        Bson::Double(v) => Some(v as i32),
+                        Bson::Int32(v) => Some(v as u32),
+                        Bson::Int64(v) => Some(v as u32),
+                        Bson::Double(v) => Some(v as u32),
                         _ => None,
                     };
                 }
@@ -55,9 +55,9 @@ impl<'de> serde::de::Visitor<'de> for ExamResultsVisitor {
                     }
 
                     number_of_questions_in_exam = match map.next_value()? {
-                        Bson::Int32(v) => Some(v),
-                        Bson::Int64(v) => Some(v as i32),
-                        Bson::Double(v) => Some(v as i32),
+                        Bson::Int32(v) => Some(v as u32),
+                        Bson::Int64(v) => Some(v as u32),
+                        Bson::Double(v) => Some(v as u32),
                         _ => None,
                     };
                 }
